Support for freebsd passwd(5)
